rename getBalanceForToken
