+++ src/components/views/donate/OneTime/SelectTokenModal/SelectTokenModal.tsx
@@ -25,7 +25,7 @@ import { shortenAddress, showToastError } from '@/lib/helpers';
 import { useGeneralWallet } from '@/providers/generalWalletProvider';
 import { wagmiConfig } from '@/wagmiConfigs';
 import { ChainType } from '@/types/config';
-import { getBalanceForToken } from './getBalanceForToken';
+import { getBalanceForToken } from './services';
 
 export interface ISelectTokenModalProps extends IModal {
 	tokens?: IProjectAcceptedToken[];
